+++ back/libs/oidc-provider/src/services/index.ts
@@ -1,2 +1,1 @@
-export * from "./oidc-provider-config-app.service";
 export * from "./oidc-provider-config.service";
